# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/pages": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get list of projects",
                "description": "Returns list of pages",
                "operationId": "7add6e88bd8d7b9dffad1c6fae578f39",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "title": "Iste quos laboris qu",
                                            "description": "<p>sdafadsfasd</p>",
                                            "image": "http://127.0.0.1:8000/storage/uploads/page/RR8liBRVijrnxWxCjYn58kEqJwCRfHsmmDzcip0j.jpg",
                                            "meta_title": "Reprehenderit dolore update",
                                            "meta_description": "Sit ullamco veniam update",
                                            "status": true
                                        }
                                    },
                                    "bool": {
                                        "summary": "A boolean value.",
                                        "value": false
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/pages/{id}": {
            "get": {
                "tags": [
                    "Pages"
                ],
                "summary": "Get pages information",
                "description": "Returns pages data",
                "operationId": "4bf369ba5de83a8a19f60a8324a4d5b4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "page id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/v1/create-page": {
            "post": {
                "tags": [
                    "Pages"
                ],
                "summary": "Create Page",
                "operationId": "3768fc24e169c5c6e4d47c72eb259bf7",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            }
                                        ]
                                    },
                                    "description": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            }
                                        ]
                                    },
                                    "meta_title": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            }
                                        ]
                                    },
                                    "meta_description": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            }
                                        ]
                                    },
                                    "keywords": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            }
                                        ]
                                    },
                                    "image": {
                                        "oneOf": [
                                            {
                                                "type": "file"
                                            }
                                        ]
                                    },
                                    "status": {
                                        "oneOf": [
                                            {
                                                "type": "boolean"
                                            }
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "page title",
                                    "slug": "slug",
                                    "description": "this is description",
                                    "meta_title": "meta title",
                                    "meta_description": "this is meta description",
                                    "keywords": "meta",
                                    "image": "this is image",
                                    "status": true
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true
                                        }
                                    },
                                    "bool": {
                                        "summary": "A boolean value.",
                                        "value": false
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/update-pages/{id}": {
            "put": {
                "tags": [
                    "Pages"
                ],
                "summary": "Update existing page",
                "operationId": "updateProject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the post to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "schema": "UpdatePostRequest",
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "meta_title": {
                                        "type": "string"
                                    },
                                    "meta_description": {
                                        "type": "string"
                                    },
                                    "keywords": {
                                        "type": "string"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            },
                            "example": {
                                "title": "page title",
                                "slug": "slug",
                                "description": "this is description",
                                "meta_title": "meta title",
                                "meta_description": "this is meta description",
                                "keywords": "meta",
                                "image": "this is image",
                                "status": true
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "success": true
                                        }
                                    },
                                    "bool": {
                                        "summary": "A boolean value.",
                                        "value": false
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/delete-page/{id}": {
            "delete": {
                "tags": [
                    "Pages"
                ],
                "summary": "Delete existing page",
                "description": "Deletes a record and returns no content",
                "operationId": "deletePage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Page id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Frontend login"
                ],
                "summary": "Login Frontend Users",
                "operationId": "30a0eb4658b2e4cd1b52ae3ec7c596e3",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "clientId": {
                                        "type": "string"
                                    },
                                    "clientSecret": {
                                        "oneOf": [
                                            {
                                                "type": "string"
                                            },
                                            {
                                                "type": "integer"
                                            }
                                        ]
                                    },
                                    "email": {
                                        "oneOf": [
                                            {
                                                "type": "email"
                                            }
                                        ]
                                    },
                                    "password": {
                                        "oneOf": [
                                            {
                                                "type": "password"
                                            }
                                        ]
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "clientId": 1,
                                    "clientSecret": 12345678,
                                    "email": "test",
                                    "password": 12345678,
                                    "grantType": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "tokenType": "Bearer",
                                            "expiresIn": 1296000,
                                            "accessToken": "access token",
                                            "refreshToken": "refresh token"
                                        }
                                    },
                                    "bool": {
                                        "summary": "A boolean value.",
                                        "value": false
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "type": "object"
                                        }
                                    ]
                                },
                                "examples": {
                                    "result": {
                                        "summary": "An result object.",
                                        "value": {
                                            "title": "The client id field is required.",
                                            "detail": "The client id field is required."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}